Refine scope-choices used by SFD grammar
This keeps them more consistent with the ones used by AFM, and also adds
better contrast with the (current) colour choices used on GitHub.

NOTE: This grammar could really do with a thorough rewrite, at the level
of detail used for the Adobe Font Metrics grammar. At the moment, values
are highlighted with hit-and-miss patterns, which is pretty sloppy work.

Typical mid-2016 rush-job.